+++ raiden-dapp/CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#2098] Input fields disabled on transfer screen when no channels are open
 - [#1838] Fixes Disclaimer mobile layout
 - [#2096] Fixes buggy wallet connection procedure
+- [#2144] Fixes navigation to transfer screen when token was selected
 
 ### Added
 - [#1941] Notification for opening channels
@@ -19,6 +20,7 @@
 [#2098]: https://github.com/raiden-network/light-client/issues/2098
 [#1838]: https://github.com/raiden-network/light-client/issues/1838
 [#2096]: https://github.com/raiden-network/light-client/issues/1838
+[#2144]: https://github.com/raiden-network/light-client/issues/2144
 
 ## [0.11.1] - 2020-08-18
 ### Fixed